requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@4150ac71867e59bb3dae938b3854c7ecb181fe6f
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.3.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
